pyproject.toml
[project]
name = "sca"
version = "0.1.0"
description = ""
authors = [
    {name = "douglas.guimaraes.2120@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3 (>=1.38.39,<2.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "google-cloud-compute (>=1.31.0,<2.0.0)",
    "google-cloud-storage (>=3.1.1,<4.0.0)",
    "google-cloud-iam (>=2.19.1,<3.0.0)",
    "google-cloud-container (>=2.57.0,<3.0.0)",
    "google-api-python-client (>=2.173.0,<3.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
